Move connectionProtocol property to shared base
